Support off-heap store and distributed caching.
Ruby objects stored in Ehcache will be automatically marshaled to a Java byte
array if the cache they are being placed in is clustered or the off-heap store
is in use. On retrieval, such Ruby objects are automatically unmarshaled before
being returned to the client. This enables the use of BigMemory and Terracotta
clustered Ehcache.